Create Fibonacci Sequence Generator in C
This program takes a starting number (st) and an ending number (et) from the user and tries to generate a sequence where each number is the sum of the previous two. It's a basic implementation that helps understand loops and number sequences in C.

When to Use:
If you're learning how loops work in C.
If you want to experiment with number sequences.
If you're just curious to see how a Fibonacci-like sequence is generated!