Create Leap Year Checker
This program checks whether a given year is a leap year or not. It follows the standard leap year rules:

A year is a leap year if it is divisible by 4.
However, if the year is also divisible by 100, it must also be divisible by 400 to be a leap year.
The program takes an input year and determines whether it is a leap year based on these conditions.